Add records for web site
Add records pointing to files02.openstack.org for zuul-ci.org and
www.zuul-ci.org.

We could make the www records a CNAME, but we can't do so for the
zone itself, because CNAME can't be used with other record types
and the zone must also carry SOA and NS records.

Change-Id: Ia7d2257876e636042b12f9f87c82772fbdb3abc3